requirements.txt
pandas
transformers
requests
dotenv
streamlit
plotly
wordcloud
matplotlib
beautifulsoup4
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
